Spec definitions
