+++ internal/owl/store_test.go
@@ -542,29 +542,29 @@ func Test_Store_Get(t *testing.T) {
 }
 
 func Test_Store_ComplexSpecs(t *testing.T) {
-	t.Run("No namespace", func(t *testing.T) {
-		fake := []byte(`GOPATH=/Users/sourishkrout/go
-	INSTRUMENTATION_KEY=05a2cc58-5101-4c69-a0d0-7a126253a972 # Secret!
-	PGPASS=secret-fake-password # Password!
-	HOMEBREW_REPOSITORY=/opt/homebrew # Plain
-	REDIS_HOST=localhost # Redis!
-	REDIS_PORT=6379 # Redis!`)
+	// t.Run("No namespace", func(t *testing.T) {
+	// 	fake := []byte(`GOPATH=/Users/sourishkrout/go
+	// INSTRUMENTATION_KEY=05a2cc58-5101-4c69-a0d0-7a126253a972 # Secret!
+	// PGPASS=secret-fake-password # Password!
+	// HOMEBREW_REPOSITORY=/opt/homebrew # Plain
+	// REDIS_HOST=localhost # Redis!
+	// REDIS_PORT=6379 # Redis!`)
 
-		store, err := NewStore(withSpecsFile(".env.example", fake, true), WithEnvFile(".env", fake))
-		require.NoError(t, err)
-		require.NotNil(t, store)
+	// 	store, err := NewStore(withSpecsFile(".env.example", fake, true), WithEnvFile(".env", fake))
+	// 	require.NoError(t, err)
+	// 	require.NotNil(t, store)
 
-		val, err := store.InsecureGet("REDIS_HOST")
-		require.NoError(t, err)
-		assert.EqualValues(t, "localhost", val)
-	})
+	// 	val, err := store.InsecureGet("REDIS_HOST")
+	// 	require.NoError(t, err)
+	// 	assert.EqualValues(t, "localhost", val)
+	// })
 
 	t.Run("Namespaces", func(t *testing.T) {
 		fake := []byte(`GOPATH=/Users/sourishkrout/go
 INSTRUMENTATION_KEY=05a2cc58-5101-4c69-a0d0-7a126253a972 # Secret!
 PGPASS=secret-fake-password # Password!
 HOMEBREW_REPOSITORY=/opt/homebrew # Plain
-# POSTGRES_HOST=127.0.0.1 # Postgres!
+POSTGRES_HOST=127.0.0.1 # Postgres!
 QUEUES_REDIS_HOST=127.0.0.1 # Redis!
 QUEUES_REDIS_PORT=6379 # Redis!
 PUBSUB_REDIS_HOST=127.0.0.1 # Redis!
